refactor(cocos-client): extract top-level code from VeilRoot into root/ (part of #1561)
Establish the root/ subdirectory scaffolding called for in #1561 and extract
the code that lives outside the Cocos @ccclass component:

  root/constants.ts      - node names, default tile sizes, feedback durations,
                           force-upgrade message
  root/types.ts          - BattleSettlementSnapshot, TutorialCampaignGuidance,
                           GlobalErrorBoundaryEvent
  root/runtime.ts        - VeilRootRuntime interface + defaults, resolver,
                           test override helpers (setVeilRootRuntimeForTests,
                           resetVeilRootRuntimeForTests)
  root/formatters.ts     - formatHeroStatBonus, formatResourceKindLabel,
                           formatUpgradeCostLabel (lifted from the class)
  root/session-helpers.ts - cloneSessionUpdate, collapseAdjacentEntries
  root/index.ts          - barrel

VeilRoot.ts imports these back so the scene-file reference is unchanged, and
re-exports the test helpers so existing test imports keep working.

This is part 1 of a multi-PR effort. The @ccclass VeilRoot component (~6600
lines, 180+ methods, nearly all bound to `this`) cannot be split further in
one PR without reshaping the class; the follow-up PRs will target
session-lifecycle, render-state-composer, panel-orchestration,
tutorial-orchestrator, prefetch-scheduler, and telemetry-hooks as described
in the issue.

No runtime behavior change. Cocos-client typecheck shows no new errors vs.
main (one pre-existing `experiments` error on profile reproduces on main at
a different line). Test results match the main baseline exactly.

Refs #1561
